check for exe exists before calling to delete

+++ beginnerApp/compile.sh
@@ -1,3 +1,6 @@
-rm main.exe
+if [ -e "main.exe" ]
+then
+    rm main.exe
+fi
 mcs main.cs
 mono main.exe
